Remove default region
Allow the default region to be determined by botocore.

Boto supports loading configuration and credentials from various
providers with a deined order of preference (e.g. environment variables,
config files, metadata service, etc). By hardcoding the defualt region
as us-east-1, we were preventing boto from looking up the correct region
dynamically as one might expect when using the AWS CLI or SDKs.

This change ensures that region configuration is aligned with AWS SDKs
and tables are not unexpectedly created in the wrong region because the
default was not overriden.

This is a breaking change!

Fixes #211