create_table should include polymorphic indexes (#1107)
With polymorphism, an index's key attribute might be defined only in a derived model (e.g. if it pertains only to a certain kind of model — for other models, it'd be a sparse index), and so it makes sense to add that index only to the derived model, e.g.
```python
class BaseModel(Model):
    ...
    cls = DescriminatorAttribute()

class DerivedIndex(GlobalSecondaryIndex):
    class Meta:
        index_name = 'derived'
        ...
    ham = UnicodeAttribute(hash_key=True)

class DerivedModel(BaseModel, discriminator='spam'):
    ham  = UnicodeAttribute()
    index = DerivedIndex()
```

A common pattern in a development environment is to use `Model.create_table`. For a polymorphic model, given you have to pick _one_ model to invoke `create_table` on, it only serves right that you invoke it on the base model.

Before this change, the base model's "schema"  didn't include all the derived models' indexes. With this change, it will.