Replace to_dict with to_simple_dict, to_dynamodb_dict (#1126)
Since #1112, `Model.serialize` return value is not necessarily JSON-serializable. We're adding `to_dynamodb_dict` as a replacement (basically, `bytes` are base-64 encoded), and `to_simple_dict` to provide for a common ask of a "simple" JSON representation similar to what the AWS Console defaults to (previously called `to_dict` in earlier versions of the 6.x branch).

We're making those methods of `AttributeContainer` rather than `Model` so they'd be equally applicable to `MapAttribute`.