feat: allow setting or unsetting the boto retry configuration (#1271)
This adds the ability to directly set the boto retry configuration
dictionary, or to leave it unset and allow botocore to automatically
discover the configuration from the environment or `~/.aws/config`
files.

The default is to use the previous PynamoDB behavior for configuring
retries so as to not make this a breaking change.